frontend/added support for multipolygon
